Backend/CybricHQ/Backend/CybricHQ/requirements.txt
# Django and REST Framework
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.37.0
django-cors-headers==4.9.0
dj-database-url==3.0.1
drf-spectacular==0.29.0

# ASGI and WebSocket Support
asgiref==3.11.0
channels==4.3.2
daphne==4.2.1

# Celery and Message Queue
celery==5.6.0
redis==7.1.0
kombu==5.6.1
amqp==5.3.1
billiard==4.2.4
vine==5.1.0

# Database
psycopg==3.3.2
psycopg-binary==3.3.2
sqlparse==0.5.4

# Async / WebSocket Libraries
Twisted==25.5.0
autobahn==25.12.2
txaio==25.12.2
Automat==25.4.16
constantly==23.10.4
hyperlink==21.0.0
Incremental==24.11.0
zope.interface==8.1.1

# Security and Cryptography
cryptography==46.0.3
pyOpenSSL==25.3.0
PyJWT==2.10.1
cffi==2.0.0
pycparser==2.23
pyasn1==0.6.1
pyasn1_modules==0.4.2
service-identity==24.2.0

# HTTP and Networking
requests==2.32.5
urllib3==2.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11

# Utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.2.14
six==1.17.0
attrs==25.4.0
packaging==25.0
typing_extensions==4.15.0
exceptiongroup==1.3.1

# JSON and Serialization
ujson==5.11.0
msgpack==1.1.2
cbor2==5.7.1
py-ubjson==0.16.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
inflection==0.5.1
uritemplate==4.2.0

# Timezone
tzdata==2025.3
tzlocal==5.3.1

# Image Processing
Pillow>=10.0.0

# Production Server
gunicorn>=21.0.0
